Got upvalues to work partially
Tracking and resolving of upvalues during compilation now appears to work,
capturing and closing at runtime appears to work, and OpCode::GetUpvalue*
tests are successful.

OpCode::SetUpvalue* does not work, because the VM is unable to get a
mutable reference to the current closure. This is probably because
a copy of the Rc object also exists on the stack (as it must).

I may have to change the Value::Obj variant from Rc<Obj> to Rc<RefCell<Obj>>
Another option may be to have the VM always access the closure through
stack slot zero, rather than using an owned copy.