Added class method declarations
Also added a "debug" statement, this is similar to the "print" statement
that shows the internal representation of a Value using std::fmt::Display,
but the debug statement uses the more verbose std::fmt::Debug to
dig into the internals of the Value. This is useful for verifying that a
Value::Obj::Class has the expected content.

Both "debug" and "print" are temporary ofcourse, in the final rlox they
should probably be replaced by something that prints numbers and strings
the way an end user would expect them to.