Can now call class methods on an instance
Bound methods are implemented in the form of a Method object,
which can be wrapped in a Value::Obj::Method so it can be called.

Also reduced debug output from the parser when compiling methods.

To simplify things, the has()/get() methods on Class and Instance
were replaced with a get() that returns Option<&Value> to
simplify things in the VM. Less borrowing, less dereferencing,
and we avoid searching the HashMap twice. And the code is more idiomatic.